[platform] Extract shared Helmert/TM projection into geodesy.hpp
os_grid_utils and irish_grid_utils each carried their own copy of the same
geodesy: the 7-parameter Helmert datum transform and the Lee-Redfearn
transverse Mercator projection. Only the constants (ellipsoid, datum shift,
projection origin) and the grid-lettering/range rules actually differ.

Move the shared machinery into a new header-only platform/geodesy.hpp,
parameterised by Ellipsoid / HelmertParams / TMParams, and reduce each grid
file to its own constants plus its lettering and region rules. This removes
~150 lines of duplicated maths and gives a single place to fix or extend the
projection.

Behaviour-preserving: the irish_grid file was already parameterised, so the
shared code is its generalisation; the OS Grid golden values pin the result.
Header-only avoids adding a new translation unit to CMake and the Xcode project.
