+++ libs/platform/irish_grid_utils.cpp
@@ -1,46 +1,28 @@
 #include "platform/irish_grid_utils.hpp"
 
+#include "platform/geodesy.hpp"
+
 #include "base/math.hpp"
 #include "base/string_utils.hpp"
 
 #include <cmath>
+#include <optional>
 #include <string>
 #include <string_view>
 
 namespace irish_grid_utils
 {
 using math::DegToRad;
-using math::RadToDeg;
 
 namespace
 {
-struct Ellipsoid
-{
-  double a;  // Semi-major axis, m.
-  double b;  // Semi-minor axis, m.
-  double constexpr EccentricitySq() const { return (a * a - b * b) / (a * a); }
-  double constexpr ThirdFlattening() const { return (a - b) / (a + b); }
-};
-
-// WGS84 (also used for ITM: ETRS89/GRS80 equals WGS84 to sub-millimetre here) and Airy Modified 1849
-// (the Ireland 1965 / TM65 ellipsoid behind the Irish Grid).
-Ellipsoid constexpr kWgs84{6378137.0, 6356752.314245};
-Ellipsoid constexpr kAiryModified{6377340.189, 6356034.447};
-
-// A transverse Mercator definition: ellipsoid + projection origin, scale and false origin.
-struct TMParams
-{
-  Ellipsoid ellipsoid;
-  double F0;    // Scale factor on the central meridian.
-  double lat0;  // True origin latitude, rad.
-  double lon0;  // True origin longitude, rad.
-  double E0;    // False easting, m.
-  double N0;    // False northing, m.
-};
+// Airy Modified 1849 (the Ireland 1965 / TM65 ellipsoid behind the Irish Grid). ITM is on ETRS89,
+// which equals WGS84 to sub-millimetre here, so it reuses geodesy::kWgs84.
+geodesy::Ellipsoid constexpr kAiryModified{6377340.189, 6356034.447};
 
 // Irish Grid (TM65) and Irish Transverse Mercator (ITM, ETRS89). Same origin, different scale/ellipsoid.
-TMParams constexpr kIrishGrid{kAiryModified, 1.000035, DegToRad(53.5), DegToRad(-8.0), 200000.0, 250000.0};
-TMParams constexpr kITM{kWgs84, 0.99982, DegToRad(53.5), DegToRad(-8.0), 600000.0, 750000.0};
+geodesy::TMParams constexpr kIrishGrid{kAiryModified, 1.000035, DegToRad(53.5), DegToRad(-8.0), 200000.0, 250000.0};
+geodesy::TMParams constexpr kITM{geodesy::kWgs84, 0.99982, DegToRad(53.5), DegToRad(-8.0), 600000.0, 750000.0};
 
 // Irish Grid lettered area: a 5x5 array of 100 km squares, i.e. [0, 500 km) on each axis.
 double constexpr kIrishGridMax = 500000.0;
@@ -52,213 +34,24 @@ bool InIreland(ms::LatLon const & ll)
   return ll.m_lat >= 51.0 && ll.m_lat <= 55.7 && ll.m_lon >= -11.0 && ll.m_lon <= -5.0;
 }
 
-struct Cartesian
-{
-  double x, y, z;
-};
-
-// 7-parameter Helmert transform. Rotations are stored in radians, scale as a ratio.
-struct HelmertParams
-{
-  double tx, ty, tz;  // Translations, m.
-  double rx, ry, rz;  // Rotations, rad.
-  double s;           // Scale, ratio.
-};
-
-double constexpr SecToRad(double sec)
-{
-  return DegToRad(sec / 3600.0);
-}
-
-HelmertParams constexpr InverseHelmert(HelmertParams const & h)
-{
-  return {-h.tx, -h.ty, -h.tz, -h.rx, -h.ry, -h.rz, -h.s};
-}
-
 // WGS84 -> Ireland 1965 (TM65). Source: OSi/OSNI "Making maps compatible with GPS" (accurate to ~1-2 m).
-HelmertParams constexpr kWgs84ToTm65{-482.530,         130.596,          -564.557, SecToRad(-1.042),
-                                     SecToRad(-0.214), SecToRad(-0.631), -8.150e-6};
-// TM65 -> WGS84 is the inverse. Negating the linearized parameters is the standard approximation
-// (round-trip error < 1 cm, far below the ~2 m Helmert accuracy).
-HelmertParams constexpr kTm65ToWgs84 = InverseHelmert(kWgs84ToTm65);
-
-Cartesian GeodeticToCartesian(ms::LatLon const & ll, Ellipsoid const & e)
-{
-  double const phi = DegToRad(ll.m_lat);
-  double const lambda = DegToRad(ll.m_lon);
-  double const sinPhi = std::sin(phi);
-  double const cosPhi = std::cos(phi);
-  double const eSq = e.EccentricitySq();
-  double const nu = e.a / std::sqrt(1.0 - eSq * sinPhi * sinPhi);  // Height is assumed zero.
-
-  return {nu * cosPhi * std::cos(lambda), nu * cosPhi * std::sin(lambda), (1.0 - eSq) * nu * sinPhi};
-}
-
-ms::LatLon CartesianToGeodetic(Cartesian const & c, Ellipsoid const & e)
-{
-  double const eSq = e.EccentricitySq();
-  double const p = std::sqrt(c.x * c.x + c.y * c.y);
-  double phi = std::atan2(c.z, p * (1.0 - eSq));
-
-  // Iterate to convergence (latitude depends on the radius of curvature, which depends on latitude).
-  for (int i = 0; i < 10; ++i)
-  {
-    double const sinPhi = std::sin(phi);
-    double const nu = e.a / std::sqrt(1.0 - eSq * sinPhi * sinPhi);
-    double const phiNew = std::atan2(c.z + eSq * nu * sinPhi, p);
-    if (std::fabs(phiNew - phi) < 1e-12)
-    {
-      phi = phiNew;
-      break;
-    }
-    phi = phiNew;
-  }
-
-  return {RadToDeg(phi), RadToDeg(std::atan2(c.y, c.x))};
-}
-
-Cartesian ApplyHelmert(Cartesian const & c, HelmertParams const & h)
-{
-  double const scale = 1.0 + h.s;
-  return {h.tx + scale * c.x - h.rz * c.y + h.ry * c.z, h.ty + h.rz * c.x + scale * c.y - h.rx * c.z,
-          h.tz - h.ry * c.x + h.rx * c.y + scale * c.z};
-}
+geodesy::HelmertParams constexpr kWgs84ToTm65{
+    -482.530, 130.596, -564.557, geodesy::SecToRad(-1.042), geodesy::SecToRad(-0.214), geodesy::SecToRad(-0.631),
+    -8.150e-6};
+geodesy::HelmertParams constexpr kTm65ToWgs84 = geodesy::InverseHelmert(kWgs84ToTm65);
 
 ms::LatLon Wgs84ToTm65(ms::LatLon const & wgs)
 {
-  return CartesianToGeodetic(ApplyHelmert(GeodeticToCartesian(wgs, kWgs84), kWgs84ToTm65), kAiryModified);
+  return geodesy::ShiftDatum(wgs, geodesy::kWgs84, kWgs84ToTm65, kAiryModified);
 }
 
 ms::LatLon Tm65ToWgs84(ms::LatLon const & tm65)
 {
-  return CartesianToGeodetic(ApplyHelmert(GeodeticToCartesian(tm65, kAiryModified), kTm65ToWgs84), kWgs84);
-}
-
-struct EN
-{
-  double easting, northing;
-};
-
-// Meridional arc from the true origin to latitude phi for the given projection. Used by both directions.
-double MeridionalArc(double phi, TMParams const & tm)
-{
-  double const n = tm.ellipsoid.ThirdFlattening();
-  double const n2 = n * n;
-  double const n3 = n2 * n;
-  double const dMinus = phi - tm.lat0;
-  double const dPlus = phi + tm.lat0;
-
-  return tm.ellipsoid.b * tm.F0 *
-         ((1.0 + n + 5.0 / 4.0 * n2 + 5.0 / 4.0 * n3) * dMinus -
-          (3.0 * n + 3.0 * n2 + 21.0 / 8.0 * n3) * std::sin(dMinus) * std::cos(dPlus) +
-          (15.0 / 8.0 * n2 + 15.0 / 8.0 * n3) * std::sin(2.0 * dMinus) * std::cos(2.0 * dPlus) -
-          35.0 / 24.0 * n3 * std::sin(3.0 * dMinus) * std::cos(3.0 * dPlus));
-}
-
-// Radii of curvature (x scale factor) and eta^2 at latitude phi. Shared by both projection directions.
-struct Curvature
-{
-  double nu;    // Transverse radius of curvature.
-  double rho;   // Meridional radius of curvature.
-  double eta2;  // nu / rho - 1.
-};
-
-Curvature ComputeCurvature(double sinPhi, TMParams const & tm)
-{
-  double const eSq = tm.ellipsoid.EccentricitySq();
-  double const t = 1.0 - eSq * sinPhi * sinPhi;
-  double const sqrtT = std::sqrt(t);
-  double const nu = tm.ellipsoid.a * tm.F0 / sqrtT;
-  double const rho = tm.ellipsoid.a * tm.F0 * (1.0 - eSq) / (t * sqrtT);
-  return {nu, rho, nu / rho - 1.0};
-}
-
-// lat/lon (on the projection's own datum) -> easting/northing (Lee-Redfearn transverse Mercator).
-EN LatLonToEN(ms::LatLon const & ll, TMParams const & tm)
-{
-  double const phi = DegToRad(ll.m_lat);
-  double const lambda = DegToRad(ll.m_lon);
-  double const sinPhi = std::sin(phi);
-  double const cosPhi = std::cos(phi);
-  double const tanPhi = sinPhi / cosPhi;
-  double const tan2 = tanPhi * tanPhi;
-  double const tan4 = tan2 * tan2;
-
-  auto const [nu, rho, eta2] = ComputeCurvature(sinPhi, tm);
-
-  double const cos3 = cosPhi * cosPhi * cosPhi;
-  double const cos5 = cos3 * cosPhi * cosPhi;
-
-  double const I = MeridionalArc(phi, tm) + tm.N0;
-  double const II = nu / 2.0 * sinPhi * cosPhi;
-  double const III = nu / 24.0 * sinPhi * cos3 * (5.0 - tan2 + 9.0 * eta2);
-  double const IIIA = nu / 720.0 * sinPhi * cos5 * (61.0 - 58.0 * tan2 + tan4);
-  double const IV = nu * cosPhi;
-  double const V = nu / 6.0 * cos3 * (nu / rho - tan2);
-  double const VI = nu / 120.0 * cos5 * (5.0 - 18.0 * tan2 + tan4 + 14.0 * eta2 - 58.0 * tan2 * eta2);
-
-  double const dLon = lambda - tm.lon0;
-  double const dLon2 = dLon * dLon;
-  double const dLon3 = dLon2 * dLon;
-  double const dLon4 = dLon2 * dLon2;
-  double const dLon5 = dLon4 * dLon;
-  double const dLon6 = dLon3 * dLon3;
-
-  double const northing = I + II * dLon2 + III * dLon4 + IIIA * dLon6;
-  double const easting = tm.E0 + IV * dLon + V * dLon3 + VI * dLon5;
-  return {easting, northing};
-}
-
-// easting/northing -> lat/lon on the projection's own datum (inverse Lee-Redfearn).
-ms::LatLon ENToLatLon(EN const & en, TMParams const & tm)
-{
-  // Iterate the latitude until the meridional arc matches the requested northing.
-  double phi = tm.lat0;
-  double m = 0.0;
-  do
-  {
-    phi += (en.northing - tm.N0 - m) / (tm.ellipsoid.a * tm.F0);
-    m = MeridionalArc(phi, tm);
-  }
-  while (std::fabs(en.northing - tm.N0 - m) >= 0.00001);  // 0.01 mm.
-
-  double const sinPhi = std::sin(phi);
-  double const cosPhi = std::cos(phi);
-  double const tanPhi = sinPhi / cosPhi;
-  double const tan2 = tanPhi * tanPhi;
-  double const tan4 = tan2 * tan2;
-  double const tan6 = tan4 * tan2;
-
-  auto const [nu, rho, eta2] = ComputeCurvature(sinPhi, tm);
-
-  double const nu3 = nu * nu * nu;
-  double const nu5 = nu3 * nu * nu;
-  double const nu7 = nu5 * nu * nu;
-  double const secPhi = 1.0 / cosPhi;
-
-  double const VII = tanPhi / (2.0 * rho * nu);
-  double const VIII = tanPhi / (24.0 * rho * nu3) * (5.0 + 3.0 * tan2 + eta2 - 9.0 * tan2 * eta2);
-  double const IX = tanPhi / (720.0 * rho * nu5) * (61.0 + 90.0 * tan2 + 45.0 * tan4);
-  double const X = secPhi / nu;
-  double const XI = secPhi / (6.0 * nu3) * (nu / rho + 2.0 * tan2);
-  double const XII = secPhi / (120.0 * nu5) * (5.0 + 28.0 * tan2 + 24.0 * tan4);
-  double const XIIA = secPhi / (5040.0 * nu7) * (61.0 + 662.0 * tan2 + 1320.0 * tan4 + 720.0 * tan6);
-
-  double const dE = en.easting - tm.E0;
-  double const dE2 = dE * dE;
-  double const dE3 = dE2 * dE;
-  double const dE4 = dE2 * dE2;
-  double const dE5 = dE4 * dE;
-  double const dE6 = dE3 * dE3;
-  double const dE7 = dE6 * dE;
-
-  double const lat = phi - VII * dE2 + VIII * dE4 - IX * dE6;
-  double const lon = tm.lon0 + X * dE - XI * dE3 + XII * dE5 - XIIA * dE7;
-  return {RadToDeg(lat), RadToDeg(lon)};
+  return geodesy::ShiftDatum(tm65, kAiryModified, kTm65ToWgs84, geodesy::kWgs84);
 }
 
 // easting/northing -> "O 1516 3447". Empty outside the Irish Grid lettered area.
-std::string ENToIrishGridRef(EN const & en, int figures)
+std::string ENToIrishGridRef(geodesy::EN const & en, int figures)
 {
   if (en.easting < 0.0 || en.easting >= kIrishGridMax || en.northing < 0.0 || en.northing >= kIrishGridMax)
     return {};
@@ -285,7 +78,7 @@ std::string ENToIrishGridRef(EN const & en, int figures)
 }
 
 // "O 152 345" / "O152345" -> easting/northing. Nullopt if malformed or outside the grid.
-std::optional<EN> IrishGridRefToEN(std::string_view gridRef)
+std::optional<geodesy::EN> IrishGridRefToEN(std::string_view gridRef)
 {
   std::string s;
   s.reserve(gridRef.size());
@@ -328,8 +121,8 @@ std::optional<EN> IrishGridRefToEN(std::string_view gridRef)
 
   // Scale the partial reference up to metres within the 100 km square (refers to its SW corner).
   long const multiplier = math::PowUint(10L, static_cast<uint64_t>(5 - figures));
-  return EN{static_cast<double>(e100k * 100000 + e * multiplier),
-            static_cast<double>(n100k * 100000 + nn * multiplier)};
+  return geodesy::EN{static_cast<double>(e100k * 100000 + e * multiplier),
+                     static_cast<double>(n100k * 100000 + nn * multiplier)};
 }
 }  // namespace
 
@@ -339,7 +132,7 @@ std::string FormatIrishGrid(double lat, double lon, int figures)
   if (!InIreland({lat, lon}))
     return {};
 
-  return ENToIrishGridRef(LatLonToEN(Wgs84ToTm65({lat, lon}), kIrishGrid), figures);
+  return ENToIrishGridRef(geodesy::LatLonToEN(Wgs84ToTm65({lat, lon}), kIrishGrid), figures);
 }
 
 std::optional<ms::LatLon> IrishGridToLatLon(std::string_view gridRef)
@@ -348,7 +141,7 @@ std::optional<ms::LatLon> IrishGridToLatLon(std::string_view gridRef)
   if (!en)
     return {};
 
-  ms::LatLon const ll = Tm65ToWgs84(ENToLatLon(*en, kIrishGrid));
+  ms::LatLon const ll = Tm65ToWgs84(geodesy::ENToLatLon(*en, kIrishGrid));
   // The lettered grid extends into the sea around Ireland; reject references off the island so a
   // successful parse always yields a real on-island point.
   if (!InIreland(ll))
@@ -362,7 +155,7 @@ std::string FormatITM(double lat, double lon)
     return {};
 
   // ETRS89 == WGS84 for our purposes, so no datum shift: project the WGS84 point directly.
-  EN const en = LatLonToEN({lat, lon}, kITM);
+  geodesy::EN const en = geodesy::LatLonToEN({lat, lon}, kITM);
   long const e = std::lround(en.easting);
   long const n = std::lround(en.northing);
   if (e < 0 || n < 0)
@@ -404,7 +197,8 @@ std::optional<ms::LatLon> ITMToLatLon(std::string_view itm)
   if (count != 2)
     return {};
 
-  ms::LatLon const ll = ENToLatLon(EN{static_cast<double>(values[0]), static_cast<double>(values[1])}, kITM);
+  ms::LatLon const ll =
+      geodesy::ENToLatLon(geodesy::EN{static_cast<double>(values[0]), static_cast<double>(values[1])}, kITM);
   if (!InIreland(ll))
     return {};
   return ll;
